Photo upload

+++ vhl/src/main/java/info/cinow/model/locationiq/LocationIqResult.java
@@ -38,7 +38,7 @@ public class LocationIqResult {
 
     private String type;
 
-    private float importance;
+    private double importance;
 
     private String icon;
